+++ src/stylesheets/admin.css
@@ -4,6 +4,11 @@
     background-color: #fff;
 }
 
+.admin-body {
+    height: max-content;
+    overflow-y: auto;
+}
+
 .admin-header {
     height: 10%;
     width: 100%;
@@ -26,38 +31,106 @@
 }
 
 .adim-options-cont {
-    height: 60%;
+    height: 55vh;
+    width: 100%;
     display: grid;
     grid-template-columns: 1fr 1fr;
     grid-template-rows: 1fr 1fr;
     place-items: center center;
     background-color: #fff;
     margin-top: 5%;
+    margin-bottom: 10vh;
 }
 
 .admin-option {
     display: flex;
     flex-direction: column;
     align-items: center;
-    width: 95%;
-    height: 95%;
+    width: 80%;
+    height: 90%;
     background-color: #2B3E9D;
     border-radius: 5px;
 }
 
 .plus-cont {
-    height: 30%;
+    display: flex;
+    justify-content: center;
+    height: 100%;
+    width: 20%;
 }
 
 .plus-img {
-    height: 100%;
-} 
+    height: 35px;
+    margin-top: 20%;
+}
 
 .admin-option-img-cont {
-    height: 100%;
+    display: flex;
+    align-items: center;
+    height: 75%;
     width: 100%;
 }
 
 .admin-option-img {
-    height: 200px;
+    height: 120px;
+    margin-left: 10%;
+}
+
+.admin-txt-cont {
+    height: 25%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.admin-txt-cont h2 {
+    color: #fff;
+    font-size: 4.2vw;
+    font-family: Itim, serif;
+    font-weight: 100;
+}
+
+.Admin-pros-cont {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 80vh;
+    background-color: #2b3e9d;
+}
+
+.Admin-pros-cont-title {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 10%;
+    color: #ffffff79;
+    font-size: 3vw;
+    font-family: Itim, serif;
+    font-weight: lighter;
+}
+
+.Admin-pros-body {
+    width: 96%;
+    height: 88%;
+    background-color: #4f63c2;
+}
+
+.Admin-pros-number-cont {
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    height: 15%;
+}
+
+.Admin-pros-number-cont h3 {
+    font-size: 3vw;
+    color: #fff;
+    font-family: Itim, serif;
+    font-weight: lighter;
+}
+
+.Admin-pros-pending-cont {
+    width: 100%;
+    height: 85%;
 }
